NETCURL-268: Update for test suite - letting core decide returned type. Recasting to integers in test.

+++ src/Module/Config/WrapperConfig.php
@@ -536,7 +536,7 @@ class WrapperConfig
             );
             $this->replaceOption(
                 WrapperCurlOpt::NETCURL_CURLOPT_TIMEOUT,
-                (float)ceil($timeout),
+                ceil($timeout),
                 WrapperCurlOpt::NETCURL_CURLOPT_TIMEOUT_MS
             );
         } else {
@@ -547,7 +547,7 @@ class WrapperConfig
             );
             $this->replaceOption(
                 WrapperCurlOpt::NETCURL_CURLOPT_TIMEOUT_MS,
-                (float)ceil($timeout),
+                ceil($timeout),
                 WrapperCurlOpt::NETCURL_CURLOPT_TIMEOUT
             );
         }
